Use cast to fix printfs format in 32bit mode.
instead of PRI maco

+++ prboom2/src/gl_preprocess.c
@@ -281,7 +281,7 @@ static void gld_FlatConvexCarver(int ssidx, int num, divline_t *list)
 
   if(!numedgepoints)
   {
-    if (levelinfo) fprintf(levelinfo, "All carved away: subsector %"PRIiPTR" - sector %i\n", (intptr_t)(ssec-subsectors), ssec->sector->iSectorID);
+    if (levelinfo) fprintf(levelinfo, "All carved away: subsector %lli - sector %i\n", (long long)(ssec-subsectors), ssec->sector->iSectorID);
   }
   else
   {
